Character Sprites

+++ Assets/Scripts/MenuParallax.cs
@@ -4,7 +4,7 @@ public class MenuParallax : MonoBehaviour
 {
     Vector2 StartPos;
 
-    [SerializeField] int moveModifier;
+    [SerializeField] float moveModifier;
 
     private void Start()
     {
